3.3V Header Fix
Both sides of the 3.3V header were on the same +3.3V net, this means that the copper fill on Layer 3 would have connected the header, not the jumper we plan to use. This would have likely broken one of our multimeters if we left it unchanged, and would make power measurements near impossible. The new power net on the output of the DCDC converter is VDD, which goes to one side of the 3.3V pin header, and the +3.3V net is on the other side.